refactor(bridge): InputValidator 패키지 생성후 BridgeSize, MoveInput 클래스 이동
